requirements.txt
flask-sqlalchemy
flask_marshmallow
marshmallow-sqlalchemy
request